+++ public/database.json
@@ -8,7 +8,7 @@
       "network": "rococo",
       "end": "500200",
       "timestamp": "2024-04-01T12:05:00Z",
-      "cost": 7,
+      "cost": "300000000000",
       "sellerAddress": "5G9Ndk3KGsLw84u7N32GHskvNx18F3x4t6uwR5SeUsU8s2LX",
       "buyerAddress": null,
       "lasticAddress": "5EUjh2hkQP67HajqcXx9hVVsdmg5QG6wUZPxA5Eje5RDPkCo",
@@ -22,7 +22,7 @@
       "network": "rococo",
       "end": "510300",
       "timestamp": "2024-04-02T08:10:00Z",
-      "cost": 5,
+      "cost": "500000000000",
       "sellerAddress": "5DnKl3eY5CzFtzL9xXDachTxPi6a2jThV9Ud9oT5sF4CcCxa",
       "buyerAddress": null,
       "lasticAddress": "5EUjh2hkQP67HajqcXx9hVVsdmg5QG6wUZPxA5Eje5RDPkCo",
@@ -36,7 +36,7 @@
       "network": "rococo",
       "end": "520400",
       "timestamp": "2024-04-03T10:15:00Z",
-      "cost": 3,
+      "cost": "200000000000",
       "sellerAddress": "5HkKm5nF5DXBq4jL5E3U3qTmU7riX8JhtXPSZ6Duh2h1MSPu",
       "buyerAddress": null,
       "lasticAddress": "5EUjh2hkQP67HajqcXx9hVVsdmg5QG6wUZPxA5Eje5RDPkCo",
@@ -50,7 +50,7 @@
       "network": "kusama",
       "end": "520400",
       "timestamp": "2024-04-03T10:15:00Z",
-      "cost": 3.5,
+      "cost": "300000000000",
       "sellerAddress": "5HkKm5nF5DXBq4jL5E3U3qTmU7riX8JhtXPSZ6Duh2h1MSPu",
       "buyerAddress": null,
       "lasticAddress": "5EUjh2hkQP67HajqcXx9hVVsdmg5QG6wUZPxA5Eje5RDPkCo",
